SwissText_2023/requirements.txt
aiohttp==3.8.1
datasets==1.18.3
nltk==3.6.7
pandas==1.4.0
pip-chill==1.0.1
pytorch-lightning==1.5.9
rouge-score==0.0.4
sacrebleu==2.0.0
seaborn==0.11.2
sentencepiece==0.1.96
spacy==3.2.1
torchaudio==0.10.2
torchvision==0.11.3
transformers==4.16.2
wandb==0.12.11
